requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
streamlit==1.29.0
tenacity==8.2.3
python-dotenv==1.0.0
google-generativeai==0.3.2
lxml==4.9.3
